+++ packages/server/src/ws.rs
@@ -2,7 +2,7 @@ pub mod handler;
 pub mod server;
 
 /// Connection ID.
-pub type ConnId = usize;
+pub type ConnId = u64;
 
 /// Room ID.
 pub type RoomId = String;
