+++ packages/tunnel_server/src/ws.rs
@@ -3,7 +3,7 @@ pub mod handler;
 pub mod server;
 
 /// Connection ID.
-pub type ConnId = usize;
+pub type ConnId = u64;
 
 /// Message sent to a room/client.
 pub type Msg = String;
